// src/app/api/generated/model/entrada-clinica-version-response.ts
/**
 * AKINE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Una version del contenido de una entrada clinica. Inmutable
 */
export interface EntradaClinicaVersionResponse { 
    /**
     * Texto clinico tal como quedo en esta version
     */
    cuerpo?: string;
    /**
     * Por que se escribio esta version. null en la 1
     */
    motivoEnmienda?: string;
    /**
     * Numero de version. La 1 es el original
     */
    numeroVersion?: number;
    /**
     * Instante UTC en que se escribio
     */
    registradaEn?: string;
    /**
     * Cuenta que la escribio
     */
    registradaPor?: number;
}